Rename: motionVariant 파일 constant로 이동
