+++ src/components/TodoModal/TodoModalDocs/LinkUpload/index.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent } from 'react';
 import { motion } from 'motion/react';
 import { PLACEHOLDERS } from '@/constants/Placeholders';
 import { Input } from '@/components/common/Input';
-import { todoModalVariants } from '@/utils/motionVariants';
+import { todoModalVariants } from '@/constants/motionVariants';
 import { useTodoDataStore } from '@/store/useTodoDataStore';
 
 export const LinkUpload = () => {
