+++ src/components/TodoModal/index.tsx
@@ -2,7 +2,7 @@ import { motion } from 'motion/react';
 import { TodoModalProps } from '@/types/TodoType';
 import { ModalContainer } from '@/components/common/Modal/ModalContainer';
 import { useTodoModalStore } from '@/store/useTodoModalStore';
-import { todoModalVariants } from '@/utils/motionVariants';
+import { todoModalVariants } from '@/constants/motionVariants';
 import { useTodoDataStore } from '@/store/useTodoDataStore';
 import { todoDataValidation } from '@/utils/todoDataValidation';
 import { Button } from '../common/Button/Button';
